+++ library/std/src/sys/unix/kernel_copy/tests.rs
@@ -159,8 +159,8 @@ fn bench_socket_pipe_socket_copy(b: &mut test::Bencher) {
     let local_source = local_end.clone();
     crate::thread::spawn(move || {
         loop {
-            crate::sys::fs::sendfile_splice(
-                crate::sys::fs::SpliceMode::Splice,
+            super::sendfile_splice(
+                super::SpliceMode::Splice,
                 local_source.as_raw_fd(),
                 write_end.as_raw_fd(),
                 u64::MAX,
